Refactoring the entire file structure. Yeah, I know I refactor a lot. Whatever, I'm learning this thing as I'm going :P
