+++ ID.WeatherScreen.pyproj
@@ -21,26 +21,51 @@
     <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="ConfigChangeHandler.py" />
-    <Compile Include="CurrentData.py" />
-    <Compile Include="ForecastData.py" />
-    <Compile Include="HistoryData.py" />
+    <Compile Include="config\ConfigChangeHandler.py" />
+    <Compile Include="config\ElementSettings.py" />
+    <Compile Include="config\FormattedTextElementSettings.py" />
+    <Compile Include="config\HumiditySquareSettings.py" />
+    <Compile Include="config\LoggingSettings.py" />
+    <Compile Include="config\RainForecastSettings.py" />
+    <Compile Include="config\RainSquareSettings.py" />
+    <Compile Include="config\SelectionSettings.py" />
+    <Compile Include="config\ServiceSettings.py" />
+    <Compile Include="config\ServicesSettings.py" />
+    <Compile Include="config\SettingsEnums.py" />
+    <Compile Include="config\SizeElementSettings.py" />
+    <Compile Include="config\SkyGradientSettings.py" />
+    <Compile Include="config\SquareElementSettings.py" />
+    <Compile Include="config\TextElementSettings.py" />
+    <Compile Include="config\WeatherSettings.py" />
+    <Compile Include="config\WindIndicatorSettings.py" />
+    <Compile Include="config\__init__.py" />
+    <Compile Include="core\__init__.py" />
+    <Compile Include="data\CurrentData.py" />
+    <Compile Include="data\ForecastData.py" />
+    <Compile Include="data\HistoryData.py" />
+    <Compile Include="data\__init__.py" />
     <Compile Include="ID.WeatherScreen.py" />
-    <Compile Include="SunData.py" />
-    <Compile Include="SunriseSunsetService.py" />
-    <Compile Include="WeatherAPIService.py" />
-    <Compile Include="WeatherConfig.py" />
-    <Compile Include="WeatherDisplay.py" />
-    <Compile Include="WeatherEncoder.py" />
-    <Compile Include="WeatherService.py" />
-    <Compile Include="WeatherUndergroundService.py" />
+    <Compile Include="data\SunData.py" />
+    <Compile Include="services\SunriseSunsetService.py" />
+    <Compile Include="services\WeatherAPIService.py" />
+    <Compile Include="config\WeatherConfig.py" />
+    <Compile Include="core\WeatherDisplay.py" />
+    <Compile Include="core\WeatherEncoder.py" />
+    <Compile Include="services\WeatherService.py" />
+    <Compile Include="services\WeatherUndergroundService.py" />
+    <Compile Include="services\__init__.py" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="assets\" />
     <Folder Include="assets\backgrounds\" />
     <Folder Include="assets\unprocessed\" />
+    <Folder Include="core\" />
+    <Folder Include="data\" />
+    <Folder Include="config\" />
+    <Folder Include="services\" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="config\ChatGPTSettings.py" />
     <Content Include="requirements.txt" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
