+++ config/ChatGPTSettings.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+@dataclass
+class ChatGPTSettings:
+    Key: str = ""
+    Model: str = "gpt-4o"
