+++ config/FormattedTextElementSettings.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+from typing import Optional
+
+from config.TextElementSettings import TextElementSettings
+
+@dataclass
+class FormattedTextElementSettings(TextElementSettings):
+    Format: Optional[str] = None
