+++ config/LoggingSettings.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+@dataclass
+class LoggingSettings:
+    EnableTrace: bool = False
+    EnableDebug: bool = False
+    LoggingLevel: str = "INFO"
