+++ config/ServiceSettings.py
@@ -1,0 +1,5 @@
+from dataclasses import dataclass
+
+@dataclass
+class ServiceSettings:
+    Key: str = ""
