+++ config/SkyGradientSettings.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+@dataclass
+class SkyGradientSettings:
+    Enable: bool = True
+    EnableCloud: bool = True
+    CloudHeight: int = 8
