+++ config/SquareElementSettings.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+from config.ElementSettings import ElementSettings
+
+@dataclass 
+class SquareElementSettings(ElementSettings):
+    Size: int = 0
