+++ config/TextElementSettings.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+from typing import Optional
+
+from config.ElementSettings import ElementSettings
+
+@dataclass
+class TextElementSettings(ElementSettings):
+    FillColor: Optional[str] = None
+    FontFamily: Optional[str] = None
+    FontWeight: Optional[str] = None
+    FontSize: Optional[int] = None
+    Anchor: Optional[str] = None
+    Stroke: bool = True
