+++ config/__init__.py
@@ -1,0 +1,21 @@
+from .ConfigChangeHandler import ConfigChangeHandler
+from .ElementSettings import ElementSettings
+from .FormattedTextElementSettings import FormattedTextElementSettings
+from .HumiditySquareSettings import HumiditySquareSettings
+from .LoggingSettings import LoggingSettings
+from .RainForecastSettings import RainForecastSettings
+from .RainSquareSettings import RainSquareSettings
+from .SelectionSettings import SelectionSettings
+from .ServiceSettings import ServiceSettings
+from .ServicesSettings import ServicesSettings
+from .SettingsEnums import (
+    TemperatureType, PressureType, WindType, VisibilityType,
+    PrecipitationType
+)
+from .SizeElementSettings import SizeElementSettings
+from .SkyGradientSettings import SkyGradientSettings
+from .SquareElementSettings import SquareElementSettings
+from .TextElementSettings import TextElementSettings
+from .WeatherSettings import WeatherSettings
+from .WindIndicatorSettings import WindIndicatorSettings
+from .WeatherConfig import WeatherConfig
