+++ data/ForecastData.py
@@ -1,7 +1,7 @@
 ﻿from datetime import datetime
-from dataclasses import dataclass, asdict, field
+from dataclasses import dataclass, field
 from enum import Enum
-from typing import Any, Dict, Optional, List
+from typing import Optional, List
 
 class MoonPhase(Enum):
     NewMoon = "New Moon"
