+++ services/__init__.py
@@ -1,0 +1,4 @@
+from .WeatherService import WeatherService
+from .WeatherAPIService import WeatherAPIService
+from .WeatherUndergroundService import WeatherUndergroundService
+from .SunriseSunsetService import SunriseSunsetService
